More NPC's added.
listVal param in tables removed in favour of using some maths in main melody_minstrel.lua to figure out values, as long as the cutscenes are in table in the order that they appear in the NPC's list.
Additional NPC: Beyzen, created to use this framework.
NPC's with some existing code now changed to use this framework, but they both need more work done to them to be considered complete.